cleanup and requirements
