+++ requirements.txt
@@ -25,3 +25,4 @@ outlines >= 0.0.27
 bitsandbytes >= 0.41.0
 loguru
 hf_transfer # for faster downloads
+sentence-transformers # for embeddings
